Move logic to Rapidez Facade
