+++ src/RapidezServiceProvider.php
@@ -10,7 +10,7 @@ use Rapidez\Core\Commands\IndexProductsCommand;
 use Rapidez\Core\Commands\InstallCommand;
 use Rapidez\Core\Commands\InstallTestsCommand;
 use Rapidez\Core\Commands\ValidateCommand;
-use Rapidez\Core\Facades\FallbackRoutes;
+use Rapidez\Core\Facades\Rapidez as RapidezFacade;
 use Rapidez\Core\Http\Controllers\Fallback\CmsPageController;
 use Rapidez\Core\Http\Controllers\Fallback\LegacyFallbackController;
 use Rapidez\Core\Http\Controllers\Fallback\UrlRewriteController;
@@ -84,9 +84,9 @@ class RapidezServiceProvider extends ServiceProvider
             $this->loadRoutesFrom(__DIR__.'/../routes/api.php');
         }
 
-        FallbackRoutes::add(UrlRewriteController::class, 5);
-        FallbackRoutes::add(CmsPageController::class, 10);
-        FallbackRoutes::add(LegacyFallbackController::class, 10000000);
+        RapidezFacade::addFallbackRoute(UrlRewriteController::class, 5);
+        RapidezFacade::addFallbackRoute(CmsPageController::class, 10);
+        RapidezFacade::addFallbackRoute(LegacyFallbackController::class, 99999);
 
         return $this;
     }
@@ -163,9 +163,8 @@ class RapidezServiceProvider extends ServiceProvider
 
     protected function registerBindings(): self
     {
-        $this->app->bind('rapidez', Rapidez::class);
+        $this->app->singleton('rapidez', Rapidez::class);
         $this->app->bind('widget-directive', WidgetDirective::class);
-        $this->app->singleton(FallbackRoutesRepository::class);
 
         return $this;
     }
